Cleanup of TextHelper
