+++ spyder/plugins/editor/utils/editor.py
@@ -26,9 +26,6 @@ from qtpy.QtGui import (QColor, QTextBlockUserData, QTextCursor, QTextBlock,
                         QTextDocument, QCursor)
 from qtpy.QtWidgets import QApplication
 
-# Local imports
-from spyder.py3compat import to_text_string
-
 
 def drift_color(base_color, factor=110):
     """
@@ -156,60 +153,6 @@ class TextHelper(object):
         except TypeError:
             self._editor_ref = editor
 
-    def goto_line(self, line, column=0, end_column=0, move=True, word=''):
-        """
-        Moves the text cursor to the specified position.
-
-        :param line: Number of the line to go to (0 based)
-        :param column: Optional column number. Default is 0 (start of line).
-        :param move: True to move the cursor. False will return the cursor
-                     without setting it on the editor.
-        :param word: Highlight the word, when moving to the line.
-        :return: The new text cursor
-        :rtype: QtGui.QTextCursor
-        """
-        line = min(line, self.line_count())
-        text_cursor = self._move_cursor_to(line)
-        if column:
-            text_cursor.movePosition(text_cursor.Right, text_cursor.MoveAnchor,
-                                     column)
-        if end_column:
-            text_cursor.movePosition(text_cursor.Right, text_cursor.KeepAnchor,
-                                     end_column)
-        if move:
-            block = text_cursor.block()
-            self.unfold_if_colapsed(block)
-            self._editor.setTextCursor(text_cursor)
-
-            if self._editor.isVisible():
-                self._editor.centerCursor()
-            else:
-                self._editor.focus_in.connect(
-                    self._editor.center_cursor_on_next_focus)
-            if word and to_text_string(word) in to_text_string(block.text()):
-                self._editor.find(word, QTextDocument.FindCaseSensitively)
-        return text_cursor
-
-    def unfold_if_colapsed(self, block):
-        """Unfold parent fold trigger if the block is collapsed.
-
-        :param block: Block to unfold.
-        """
-        try:
-            folding_panel = self._editor.panels.get('FoldingPanel')
-        except KeyError:
-            pass
-        else:
-            from spyder.plugins.editor.utils.folding import FoldScope
-            if not block.isVisible():
-                block = FoldScope.find_parent_scope(block)
-                if TextBlockHelper.is_collapsed(block):
-                    folding_panel.toggle_fold_trigger(block)
-
-    def selected_text(self):
-        """Returns the selected text."""
-        return self._editor.textCursor().selectedText()
-
     def word_under_cursor(self, select_whole_word=False, text_cursor=None):
         """
         Gets the word under cursor using the separators defined by
@@ -280,40 +223,6 @@ class TextHelper(object):
         text_cursor = self.word_under_cursor(True, text_cursor)
         return text_cursor
 
-    def cursor_position(self):
-        """
-        Returns the QTextCursor position. The position is a tuple made up of
-        the line number (0 based) and the column number (0 based).
-
-        :return: tuple(line, column)
-        """
-        return (self._editor.textCursor().blockNumber(),
-                self._editor.textCursor().columnNumber())
-
-    def current_line_nbr(self):
-        """
-        Returns the text cursor's line number.
-
-        :return: Line number
-        """
-        return self.cursor_position()[0]
-
-    def current_column_nbr(self):
-        """
-        Returns the text cursor's column number.
-
-        :return: Column number
-        """
-        return self.cursor_position()[1]
-
-    def line_count(self):
-        """
-        Returns the line count of the specified editor.
-
-        :return: number of lines in the document.
-        """
-        return self._editor.document().blockCount()
-
     def line_text(self, line_nbr):
         """
         Gets the text of the specified line.
@@ -332,8 +241,8 @@ class TextHelper(object):
         Gets the previous line text (relative to the current cursor pos).
         :return: previous line text (str)
         """
-        if self.current_line_nbr():
-            return self.line_text(self.current_line_nbr() - 1)
+        if self._editor.current_line_nbr():
+            return self.line_text(self._editor.current_line_nbr() - 1)
         return ''
 
     def current_line_text(self):
@@ -342,7 +251,7 @@ class TextHelper(object):
 
         :return: Text of the current line
         """
-        return self.line_text(self.current_line_nbr())
+        return self.line_text(self._editor.current_line_nbr())
 
     def set_line_text(self, line_nbr, new_text):
         """
@@ -353,83 +262,10 @@ class TextHelper(object):
 
         """
         editor = self._editor
-        text_cursor = self._move_cursor_to(line_nbr)
+        text_cursor = self._editor._move_cursor_to(line_nbr)
         text_cursor.select(text_cursor.LineUnderCursor)
         text_cursor.insertText(new_text)
         editor.setTextCursor(text_cursor)
-
-    def remove_last_line(self):
-        """Removes the last line of the document."""
-        editor = self._editor
-        text_cursor = editor.textCursor()
-        text_cursor.movePosition(text_cursor.End, text_cursor.MoveAnchor)
-        text_cursor.select(text_cursor.LineUnderCursor)
-        text_cursor.removeSelectedText()
-        text_cursor.deletePreviousChar()
-        editor.setTextCursor(text_cursor)
-
-    def clean_document(self):
-        """
-        Removes trailing whitespaces and ensure one single blank line at the
-        end of the QTextDocument.
-
-        FIXME: It was deprecated in pyqode, maybe It should be deleted
-        """
-        editor = self._editor
-        value = editor.verticalScrollBar().value()
-        pos = self.cursor_position()
-        editor.textCursor().beginEditBlock()
-
-        # cleanup whitespaces
-        editor._cleaning = True
-        eaten = 0
-        removed = set()
-        for line in editor._modified_lines:
-            # parse line before and line after modified line (for cases where
-            # key_delete or key_return has been pressed)
-            for j in range(-1, 2):
-                # skip current line
-                if line + j != pos[0]:
-                    if line + j >= 0:
-                        txt = self.line_text(line + j)
-                        stxt = txt.rstrip()
-                        if txt != stxt:
-                            self.set_line_text(line + j, stxt)
-                        removed.add(line + j)
-        editor._modified_lines -= removed
-
-        # ensure there is only one blank line left at the end of the file
-        i = self.line_count()
-        while i:
-            line = self.line_text(i - 1)
-            if line.strip():
-                break
-            self.remove_last_line()
-            i -= 1
-        if self.line_text(self.line_count() - 1):
-            editor.appendPlainText('')
-
-        # restore cursor and scrollbars
-        text_cursor = editor.textCursor()
-        doc = editor.document()
-        assert isinstance(doc, QTextDocument)
-        text_cursor = self._move_cursor_to(pos[0])
-        text_cursor.movePosition(text_cursor.StartOfLine,
-                                 text_cursor.MoveAnchor)
-        cpos = text_cursor.position()
-        text_cursor.select(text_cursor.LineUnderCursor)
-        if text_cursor.selectedText():
-            text_cursor.setPosition(cpos)
-            offset = pos[1] - eaten
-            text_cursor.movePosition(text_cursor.Right, text_cursor.MoveAnchor,
-                                     offset)
-        else:
-            text_cursor.setPosition(cpos)
-        editor.setTextCursor(text_cursor)
-        editor.verticalScrollBar().setValue(value)
-
-        text_cursor.endEditBlock()
-        editor._cleaning = False
 
     def select_whole_line(self, line=None, apply_selection=True):
         """
@@ -442,57 +278,8 @@ class TextHelper(object):
         :return: QTextCursor
         """
         if line is None:
-            line = self.current_line_nbr()
+            line = self._editor.current_line_nbr()
         return self.select_lines(line, line, apply_selection=apply_selection)
-
-    def _move_cursor_to(self, line):
-        cursor = self._editor.textCursor()
-        block = self._editor.document().findBlockByNumber(line-1)
-        cursor.setPosition(block.position())
-        return cursor
-
-    def select_lines(self, start=0, end=-1, apply_selection=True):
-        """
-        Selects entire lines between start and end line numbers.
-
-        This functions apply the selection and returns the text cursor that
-        contains the selection.
-
-        Optionally it is possible to prevent the selection from being applied
-        on the code editor widget by setting ``apply_selection`` to False.
-
-        :param start: Start line number (0 based)
-        :param end: End line number (0 based). Use -1 to select up to the
-            end of the document
-        :param apply_selection: True to apply the selection before returning
-         the QTextCursor.
-        :returns: A QTextCursor that holds the requested selection
-        """
-        editor = self._editor
-        if end == -1:
-            end = self.line_count() - 1
-        if start < 0:
-            start = 0
-        text_cursor = self._move_cursor_to(start)
-        if end > start:  # Going down
-            text_cursor.movePosition(text_cursor.Down,
-                                     text_cursor.KeepAnchor, end - start)
-            text_cursor.movePosition(text_cursor.EndOfLine,
-                                     text_cursor.KeepAnchor)
-        elif end < start:  # going up
-            # don't miss end of line !
-            text_cursor.movePosition(text_cursor.EndOfLine,
-                                     text_cursor.MoveAnchor)
-            text_cursor.movePosition(text_cursor.Up,
-                                     text_cursor.KeepAnchor, start - end)
-            text_cursor.movePosition(text_cursor.StartOfLine,
-                                     text_cursor.KeepAnchor)
-        else:
-            text_cursor.movePosition(text_cursor.EndOfLine,
-                                     text_cursor.KeepAnchor)
-        if apply_selection:
-            editor.setTextCursor(text_cursor)
-        return text_cursor
 
     def selection_range(self):
         """
@@ -512,49 +299,6 @@ class TextHelper(object):
             end -= 1
         return start, end
 
-    def line_pos_from_number(self, line_number):
-        """
-        Computes line position on Y-Axis (at the center of the line) from line
-        number.
-
-        :param line_number: The line number for which we want to know the
-                            position in pixels.
-        :return: The center position of the line.
-        """
-        editor = self._editor
-        block = editor.document().findBlockByNumber(line_number)
-        if block.isValid():
-            return int(editor.blockBoundingGeometry(block).translated(
-                editor.contentOffset()).top())
-        if line_number <= 0:
-            return 0
-        else:
-            return int(editor.blockBoundingGeometry(
-                block.previous()).translated(editor.contentOffset()).bottom())
-
-    def line_nbr_from_position(self, y_pos):
-        """
-        Returns the line number from the y_pos.
-
-        :param y_pos: Y pos in the editor
-        :return: Line number (0 based), -1 if out of range
-        """
-        editor = self._editor
-        height = editor.fontMetrics().height()
-        for top, line, block in editor.visible_blocks:
-            if top <= y_pos <= top + height:
-                return line
-        return -1
-
-    def mark_whole_doc_dirty(self):
-        """
-        Marks the whole document as dirty to force a full refresh. **SLOW**
-        """
-        text_cursor = self._editor.textCursor()
-        text_cursor.select(text_cursor.Document)
-        self._editor.document().markContentsDirty(text_cursor.selectionStart(),
-                                                  text_cursor.selectionEnd())
-
     def line_indent(self, line_nbr=None):
         """
         Returns the indent level of the specified line.
@@ -566,7 +310,7 @@ class TextHelper(object):
                  current line
         """
         if line_nbr is None:
-            line_nbr = self.current_line_nbr()
+            line_nbr = self._editor.current_line_nbr()
         elif isinstance(line_nbr, QTextBlock):
             line_nbr = line_nbr.blockNumber()
         line = self.line_text(line_nbr)
@@ -618,12 +362,6 @@ class TextHelper(object):
         if keep_position:
             text_cursor.setPosition(s)
             text_cursor.setPosition(e, text_cursor.KeepAnchor)
-        self._editor.setTextCursor(text_cursor)
-
-    def clear_selection(self):
-        """Clears text cursor selection."""
-        text_cursor = self._editor.textCursor()
-        text_cursor.clearSelection()
         self._editor.setTextCursor(text_cursor)
 
     def move_right(self, keep_anchor=False, nb_chars=1):
@@ -687,40 +425,6 @@ class TextHelper(object):
             cursor.setPosition(cursor.position() + 1)
             cursor = text_document.find(search_txt, cursor, search_flags)
         return occurrences, index
-
-    def is_comment_or_string(self, cursor_or_block, formats=None):
-        """
-        Checks if a block/cursor is a string or a comment.
-        :param cursor_or_block: QTextCursor or QTextBlock
-        :param formats: the list of color scheme formats to consider. By
-            default, it will consider the following keys: 'comment', 'string',
-            'docstring'.
-        """
-        if formats is None:
-            formats = ["comment", "string", "docstring"]
-        layout = None
-        pos = 0
-        if isinstance(cursor_or_block, QTextBlock):
-            pos = len(cursor_or_block.text()) - 1
-            layout = cursor_or_block.layout()
-        elif isinstance(cursor_or_block, QTextCursor):
-            b = cursor_or_block.block()
-            pos = cursor_or_block.position() - b.position()
-            layout = b.layout()
-        if layout is not None:
-            additional_formats = layout.additionalFormats()
-            sh = self._editor.syntax_highlighter
-            if sh:
-                ref_formats = sh.color_scheme.formats
-                for r in additional_formats:
-                    if r.start <= pos < (r.start + r.length):
-                        for fmt_type in formats:
-                            is_user_obj = (r.format.objectType() ==
-                                           r.format.UserObject)
-                            if (ref_formats[fmt_type] == r.format and
-                                    is_user_obj):
-                                return True
-        return False
 
     def select_extended_word(self, continuation_chars=('.',)):
         """
@@ -1035,50 +739,6 @@ class ParenthesisInfo(object):
         self.position = pos
         #: The parenthesis character, one of "(", ")", "{", "}", "[", "]"
         self.character = char
-
-
-def get_block_symbol_data(editor, block):
-    """
-    Gets the list of ParenthesisInfo for specific text block.
-
-    :param editor: Code editor instance
-    :param block: block to parse
-    """
-    def list_symbols(editor, block, character):
-        """
-        Retuns  a list of symbols found in the block text
-
-        :param editor: code editor instance
-        :param block: block to parse
-        :param character: character to look for.
-        """
-        text = block.text()
-        symbols = []
-        cursor = QTextCursor(block)
-        cursor.movePosition(cursor.StartOfBlock)
-        pos = text.find(character, 0)
-        cursor.movePosition(cursor.Right, cursor.MoveAnchor, pos)
-
-        while pos != -1:
-            if not TextHelper(editor).is_comment_or_string(cursor):
-                # skips symbols in string literal or comment
-                info = ParenthesisInfo(pos, character)
-                symbols.append(info)
-            pos = text.find(character, pos + 1)
-            cursor.movePosition(cursor.StartOfBlock)
-            cursor.movePosition(cursor.Right, cursor.MoveAnchor, pos)
-        return symbols
-
-    parentheses = sorted(
-        list_symbols(editor, block, '(') + list_symbols(editor, block, ')'),
-        key=lambda x: x.position)
-    square_brackets = sorted(
-        list_symbols(editor, block, '[') + list_symbols(editor, block, ']'),
-        key=lambda x: x.position)
-    braces = sorted(
-        list_symbols(editor, block, '{') + list_symbols(editor, block, '}'),
-        key=lambda x: x.position)
-    return parentheses, square_brackets, braces
 
 
 def keep_tc_pos(func):
