+++ spyder/widgets/findreplace.py
@@ -26,7 +26,6 @@ from spyder.config.gui import config_shortcut
 from spyder.py3compat import to_text_string
 from spyder.utils import icon_manager as ima
 from spyder.utils.misc import regexp_error_msg
-from spyder.plugins.editor.utils.editor import TextHelper
 from spyder.utils.qthelpers import create_toolbutton, get_icon
 from spyder.widgets.comboboxes import PatternComboBox
 
@@ -424,7 +423,7 @@ class FindReplace(QWidget):
 
             if self.is_code_editor and found:
                 block = self.editor.textCursor().block()
-                TextHelper(self.editor).unfold_if_colapsed(block)
+                self.editor.unfold_if_colapsed(block)
 
                 if rehighlight or not self.editor.found_results:
                     self.highlight_timer.stop()
